docker/requirements.txt
psycopg2-binary
Faker
passlib[argon2]
